requirements.txt

celery==5.4.*
croniter==5.0.1
django==5.1.*
django-celery-beat==2.7.0
djangorestframework==3.15.*
django-rest-swagger==2.2.*
django-filter==24.3
drf-yasg==1.21.*
psycopg2==2.9.*
redis==5.1.*
setuptools==75.1.*


